# packages/backend/src/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/app-lifecycle/{urn}/restart": {
      "post": {
        "operationId": "restartApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
...
  },
  "info": {
    "title": "Runtipi API",
    "description": "API specs for Runtipi",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LifecycleRequestDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string",
                "description": "a UUID",
                "format": "uuid",
                "pattern": "[\\da-f]{8}-[\\da-f]{4}-[1-8][\\da-f]{3}-[89ab][\\da-f]{3}-[\\da-f]{12}"
              },
              {
                "const": "00000000-0000-0000-0000-000000000000",
                "description": "a UUID",
                "format": "uuid"
              },
              {
                "const": "ffffffff-ffff-ffff-ffff-ffffffffffff",
                "description": "a UUID",
                "format": "uuid"
              }
            ],
            "description": "a UUID",
            "format": "uuid"
          }
        },
        "required": [
          "requestId"
        ]
      },
...
    }
  }
}